Move all code into the common directory.
