+++ common/processors/stream_from_path.py
