--- plumbing/status.py
@@ -1,3 +1,0 @@
-"""This module updates the status of the pipeline in the README."""
-
-# TODO: implement this as a git hook
